thcrap: Improved speed of patch file resolution

+++ thcrap/src/breakpoint.cpp
@@ -220,7 +220,7 @@ static const size_t bp_entry_index = &bp_entry_indexptr + 1 - (uint8_t*)&bp_entr
 static const size_t bp_entry_local = &bp_entry_localptr + 1 - (uint8_t*)&bp_entry;
 static const size_t bp_entry_call  = &bp_entry_callptr  + 1 - (uint8_t*)&bp_entry;
 
-int breakpoints_apply(breakpoint_t *breakpoints, size_t bp_count, HMODULE hMod)
+size_t breakpoints_apply(breakpoint_t *breakpoints, size_t bp_count, HMODULE hMod)
 {
 	if(!breakpoints || !bp_count) {
 		log_print("No breakpoints to set up.\n");
